backend/requirements.txt
fastapi
uvicorn
sqlalchemy
oracledb
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
boto3
